Cleaning

+++ source/include/op_vec_sror_impl.hpp
@@ -54,7 +54,7 @@ _END_OP
 
 _DEFINE_OP(VSROR_64) {
   // Super naive reference implementation
-  uint64  shf = vm->gpr[(vArgs & 0xF000) >> 12].u8() & 0x3F;
+  uint32  shf = vm->gpr[(vArgs & 0xF000) >> 12].u8() & 0x3F;
   uint64* src = vm->gpr[(vArgs & 0x0F00) >> 8].pU64();
   uint64* dst = vm->gpr[(vArgs & 0x00F0) >> 4].pU64();
   uint32  i   = vm->gpr[(vArgs & 0x000F)].u32();
